[semantic-arc] Extract out borrow scope optimizations into its own file.
