+++ lib/SILOptimizer/SemanticARC/CMakeLists.txt
@@ -1,4 +1,5 @@
 target_sources(swiftSILOptimizer PRIVATE
   SemanticARCOpts.cpp
   OwnershipLiveRange.cpp
-  LoadCopyToLoadBorrowOpt.cpp)
+  LoadCopyToLoadBorrowOpt.cpp
+  BorrowScopeOpts.cpp)
